refactor: renames <ForgotPassword/> page-component, test files and improves its compositions
