+++ template/src/pages/forgot-password/__tests__/ForgotPassword.test.tsx
@@ -1,31 +1,35 @@
 import userEvent from '@testing-library/user-event';
-import ForgotPassword from '../index';
+import ForgotPassword from '../ForgotPassword';
 import { renderWithRouter, screen, waitFor } from '../../../utils/test';
 import { makeServer } from '../../../server';
 
-let server: any;
-
-beforeEach(() => {
-  server = makeServer();
-});
-
-afterEach(() => {
-  server.shutdown();
-});
-
 describe('<ForgotPassword />', () => {
-  it('should render the ForgotPassword form UI', () => {
+  // eslint-disable-next-line
+  let server: any;
+
+  beforeEach(() => {
+    server = makeServer();
+  });
+
+  afterEach(() => {
+    server.shutdown();
+  });
+
+  it('should render the ForgotPassword form UI', async () => {
     renderWithRouter(<ForgotPassword />);
 
-    expect(
-      screen.getByText(/Fill the field below with your email/i)
-    ).toBeInTheDocument();
-    expect(screen.getByLabelText('email')).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: /Send/i })).toBeInTheDocument();
+    await waitFor(() => {
+      expect(
+        screen.getByText(/Fill the field below with your email/i)
+      ).toBeInTheDocument();
+      expect(screen.getByLabelText('email')).toBeInTheDocument();
+      expect(screen.getByRole('button', { name: /Send/i })).toBeInTheDocument();
+    });
   });
 
   it('should render error messages when inputs are invalid', async () => {
     renderWithRouter(<ForgotPassword />);
+
     userEvent.type(screen.getByLabelText('email'), 'invalidemail');
 
     await waitFor(() => {
@@ -44,8 +48,9 @@ describe('<ForgotPassword />', () => {
 
     userEvent.type(screen.getByLabelText('email'), 'incorrect@t10.digital');
     userEvent.click(screen.getByTestId('submit-forgot'));
+
     await waitFor(() => {
-      expect(screen.getByText(/Something went wrong/i)).toBeInTheDocument();
+      expect(screen.getByText(/user not found/i)).toBeInTheDocument();
     });
   });
 });
